Tasks: - Implementing the service layer which will take care of parsing the yaml configuration and handle the download process - Unit Tests

+++ CLI.Down.App/CLI.Down.Entities/CommandParser.cs
@@ -5,6 +5,6 @@ public class CommandParser
     public bool IsVerbose { get; set; }
     public bool IsDryRun { get; set; }
     public bool IsParallel { get; set; }
-    public byte ParallelDownloads { get; set; }
+    public int ParallelDownloads { get; set; }
     public string? YamlPath { get; set; }
 }
